Refactor demoCode to shared constants module
