+++ ui/src/constants/index.ts
@@ -1,0 +1,21 @@
+export const demoCode = `/**
+ * Halo main class.
+ *
+ * @author ryanwang
+ * @author JohnNiang
+ * @author guqing
+ * @date 2017-11-14
+ */
+@EnableScheduling
+@SpringBootApplication(scanBasePackages = "run.halo.app", exclude =
+    IntegrationAutoConfiguration.class)
+@ConfigurationPropertiesScan(basePackages = "run.halo.app.infra.properties")
+public class Application {
+
+    public static void main(String[] args) {
+        new SpringApplicationBuilder(Application.class)
+            .applicationStartup(new BufferingApplicationStartup(1024))
+            .run(args);
+    }
+
+}`;
