+++ ui/src/formkit/ShikiEditorPreview.vue
@@ -1,28 +1,7 @@
 <script lang="ts" setup>
 import { VLoading } from "@halo-dev/components";
 import { ref, watchEffect } from "vue";
-
-const demoCode = `/**
- * Halo main class.
- *
- * @author ryanwang
- * @author JohnNiang
- * @author guqing
- * @date 2017-11-14
- */
-@EnableScheduling
-@SpringBootApplication(scanBasePackages = "run.halo.app", exclude =
-    IntegrationAutoConfiguration.class)
-@ConfigurationPropertiesScan(basePackages = "run.halo.app.infra.properties")
-public class Application {
-
-    public static void main(String[] args) {
-        new SpringApplicationBuilder(Application.class)
-            .applicationStartup(new BufferingApplicationStartup(1024))
-            .run(args);
-    }
-
-}`;
+import { demoCode } from "../constants";
 
 const { theme } = defineProps<{
   theme?: string;
