+++ ui/src/formkit/ShikiThemePreview.vue
@@ -2,30 +2,9 @@
 import { VTabbar } from "@halo-dev/components";
 import { useScriptTag } from "@vueuse/core";
 import { computed, onUnmounted, ref, watch } from "vue";
+import { demoCode } from "../constants";
 
 type ColorScheme = "light" | "dark";
-
-const demoCode = `/**
- * Halo main class.
- *
- * @author ryanwang
- * @author JohnNiang
- * @author guqing
- * @date 2017-11-14
- */
-@EnableScheduling
-@SpringBootApplication(scanBasePackages = "run.halo.app", exclude =
-    IntegrationAutoConfiguration.class)
-@ConfigurationPropertiesScan(basePackages = "run.halo.app.infra.properties")
-public class Application {
-
-    public static void main(String[] args) {
-        new SpringApplicationBuilder(Application.class)
-            .applicationStartup(new BufferingApplicationStartup(1024))
-            .run(args);
-    }
-
-}`;
 
 const { variant, lightTheme, darkTheme } = defineProps<{
   variant?: string;
